Initial workings making a basic 3d rotating cube.
Playing around with storage. Surfaces being an array of pointers to Points is a bit confusing but any changes to the cube verticies will show up in the surfaces so that's nice.

Just need to do some kind of z buffering or whatever. Can probably just take the average or max depth of each surface and just draw them deepest to shallowest. For a cube at least, there are no overlapping surfaces so that shouldn't be a problem doing it this way.

Also need to convert to something that will work well on the 128x128 display. Maybe remove some floats. Scaling at the end is probably okay too.

Right now just plotting in Matlab from the program output.